Terry West major cleanup to keep various Linux compilers happy, as well as fixing various bugs those compilers turned up.  Note additional compiler flags in the top-level Makefile.

+++ e19/e.cm.c
@@ -806,7 +806,7 @@ Flag on;
 		if (fileflags[curfile] & CANMODIFY)
 		    break;
 		if (fileflags[curfile] & FWRITEABLE ||
-			userid == fgetowner (curfile))
+			userid == (unsigned int)fgetowner (curfile))
 		    fileflags[curfile] |= CANMODIFY;
 		break;
